requirements.txt
numpy
mrcfile
Pillow
